requirements.txt
aws-cdk-lib==2.80.0
constructs